# go-api/handler/auth.go
package handler

import (
	"errors"
	"log"
	"net/mail"
	"time"

	"API/config"
	"API/database"
	"API/model"

	"gorm.io/gorm"

	"github.com/gofiber/fiber/v2"
	"github.com/golang-jwt/jwt/v5"
	"golang.org/x/crypto/bcrypt"
)

type LoginInput struct {
	Identity string `json:"identity" example:"user2"`
	Password string `json:"password" example:"user1234"`
}

type LoginResponse struct {
	HttpResponse
	Data LoginData
}

type LoginData struct {
	Token string `json:"token" example:"it returns jwt token"`
}

// CheckPasswordHash compare password with hash
func CheckPasswordHash(password, hash string) bool {
	t1 := time.Now()
	err := bcrypt.CompareHashAndPassword([]byte(hash), []byte(password))
	elapsed := time.Since(t1)
	log.Println("Elapsed hash compare time:", elapsed)
	return err == nil
}

func getUserByEmail(e string) (*model.User, error) {
	db := database.DB
	var user model.User
	if err := db.Where(&model.User{Email: e}).First(&user).Error; err != nil {
		if errors.Is(err, gorm.ErrRecordNotFound) {
			return nil, nil
		}
		return nil, err
	}
	return &user, nil
}

func getUserByUsername(u string) (*model.User, error) {
	db := database.DB
	var user model.User
	if err := db.Where(&model.User{Username: u}).First(&user).Error; err != nil {
		if errors.Is(err, gorm.ErrRecordNotFound) {
			return nil, nil
		}
		return nil, err
	}
	return &user, nil
}

func valid(email string) bool {
	_, err := mail.ParseAddress(email)
	return err == nil
}

// Login get user and password
//
//	@Summary		User login
//	@Description	Authenticate user by username or email and password, returns JWT token on success
//	@Tags			auth
//	@Accept			json
//	@Produce		json
//	@Param			loginInput	body		LoginInput	true	"Login credentials"
//	@Success		200			{object}	LoginResponse
//	@Failure		400			{object}	HttpResponse{}
//	@Failure		401			{object}	HttpResponse{}
//	@Failure		500			{object}	HttpResponse{}
//	@Router			/auth/login [post]
func Login(c *fiber.Ctx) error {
	type UserData struct {
		ID       uint   `json:"id"`
		Username string `json:"username"`
		Email    string `json:"email"`
		Password string `json:"password"`
	}
	input := new(LoginInput)
	var ud UserData

	if err := c.BodyParser(input); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"status": "error", "message": "Error on login request", "data": err.Error()})
	}

	identity := input.Identity
	pass := input.Password
	userModel, err := new(model.User), *new(error)

	if valid(identity) {
		userModel, err = getUserByEmail(identity)
	} else {
		userModel, err = getUserByUsername(identity)
	}

	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"status": "error", "message": "Internal Server Error", "data": err})
	} else if userModel == nil {
		CheckPasswordHash(pass, "")
		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"status": "error", "message": "Invalid identity or password", "data": err})
	} else {
		ud = UserData{
			ID:       userModel.ID,
			Username: userModel.Username,
			Email:    userModel.Email,
			Password: userModel.Password,
		}
	}

	r := CheckPasswordHash(pass, ud.Password)
	if !r {
		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"status": "error", "message": "Invalid identity or password", "data": nil})
	}

	token := jwt.New(jwt.SigningMethodHS256)

	claims := token.Claims.(jwt.MapClaims)
	claims["username"] = ud.Username
	claims["user_id"] = ud.ID
	claims["exp"] = time.Now().Add(time.Hour * 72).Unix()

	t, err := token.SignedString([]byte(config.Config("SECRET")))
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"status": "error", "message": "Internal Server Error", "data": nil})
	}

	return c.JSON(fiber.Map{"status": "success", "message": "Success login", "data": fiber.Map{"token": t}})
}